Graphics.UI.WXCore.Controls

Description

Log

textCtrlMakeLogActiveTarget :: TextCtrl a -> IO ()

Set a text control as a log target.

logDeleteAndSetActiveTarget :: Log a -> IO ()

Sets the active log target and deletes the old one.

Tree control

data TreeCookie

Represents the children of a tree control.

treeCtrlGetChildCookie :: TreeCtrl a -> TreeItem -> IO TreeCookie

Get a TreeCookie to iterate through the children of tree node.

treeCtrlGetNextChild2 :: TreeCtrl a -> TreeCookie -> IO (Maybe TreeItem)

Get the next child of a tree node. Returns Nothing when the end of the list is reached. This also invalidates the tree cookie.

treeCtrlWithChildren :: TreeCtrl a -> TreeItem -> (TreeItem -> IO b) -> IO [b]

Iterate on the list of children of a tree node.

treeCtrlGetChildren :: TreeCtrl a -> TreeItem -> IO [TreeItem]

Get the children of tree node.

treeCtrlGetSelections2 :: TreeCtrl a -> IO [TreeItem]

Get the selections of a tree control.

Wrappers

listBoxGetSelectionList :: ListBox a -> IO [Int]

Return the current selection in a listbox.

execClipBoardData :: Clipboard a -> (Clipboard a -> IO b) -> IO b

Use a clipboardSetData or clipboardGetData in this function. But don't use long computation in this function. Because this function encloses the computation with clipboardOpen and clipboardClose, and wxHaskell uses Global clipboard on your environment. So, long computation causes problem.

Font Enumerator

enumerateFontsList :: Int -> Bool -> IO [String]

(enumerateFontsList encoding fixedWidthOnly) return the Names of the available fonts in a list. To get all available fonts call enumerateFontsList wxFONTENCODING_SYSTEM False. See also enumerateFonts.

enumerateFonts :: Int -> Bool -> (String -> IO Bool) -> IO ()

(enumerateFonts encoding fixedWidthOnly f calls successive f name for the fonts installed on the system. It stops if the function return False. See also enumerateFontsList.

Deprecated

wxcAppUSleep :: Int -> IO ()

Deprecated: Use wxcAppMilliSleep instead

This function just left for backward-compatiblity. Update your code to use wxcAppMilliSleep instead.