Graphics.UI.WXCore.Defines

Description

Exports standard defines of wxWidgets.

GUI toolkit

data WxToolkit

wxWidgets library kinds.

Constructors

WxGTK
GTK

WxMac
MacOS

WxMSW
Any windows

WxMotif

WxMGL
SciTech soft MGL

WxUniversal

WxOSTwo
OS/2

WxXEleven
X11

WxUnknown

wxToolkit :: WxToolkit

Get the current wxWidgets library kind.

wxVersion :: Int

Return the version of the wxWidgets library. It is composed of the major version times 1000, plus the minor version times 100, plus the release number. For example, version 2.1.15 would be 2115.

Files

getAbsoluteFilePath :: FilePath -> IO FilePath

Find a file relative to the application or current directory. (absolute paths are passed without modification). This allows one to access resources relative to the installation directory in a portable way.

dirSep :: String

Deprecated: Use System.FilePath module and/or its module's pathSeparator instead

deprecated: Use System.FilePath module and/or its module's pathSeparator instead.

pathSep :: String

Deprecated: Use System.FilePath module's searchPathSeparator instead

deprecated: Use System.FilePath module's searchPathSeparator instead.