Graphics.UI.WXCore.DragAndDrop

Description

Drag & Drop events.

Drop Targets

dropTarget :: Window a -> DataObject b -> IO (WXCDropTarget ())

Set a drop target window and DataObject that is associated with drop event.

Create Drop Soruces

dropSource :: DataObject a -> Window b -> IO (DropSource ())

Create DropSource. Then dragAndDrop replace target's DataObject by this DataObject.

dropSourceWithCursor :: DataObject a -> Window b -> Cursor c -> Cursor c -> Cursor c -> IO (DropSource ())

On Windows or Mac OS X platform, you must choose this function or dropSourceWithCursorByString, if you want to use Custom Cursor for Drag & Drop event. dropSourceWithIcon doesn't work these platform, and this function doesn't work other platforms.

dropSourceWithCursorByString :: DataObject a -> Window b -> String -> String -> String -> IO (DropSource ())

dropSourceWithIcon :: DataObject a -> Window b -> Icon c -> Icon c -> Icon c -> IO (DropSource ())