Graphics.UI.WXCore.Frame

Description

Frame utility functions.

Frame

frameCreateTopFrame :: String -> IO (Frame ())

Create a default frame and make it the top-level window.

frameCreateDefault :: String -> IO (Frame ())

Create a frame with default settings.

frameSetTopFrame :: Frame a -> IO ()

Set the top-level frame (calls cAppSetTopWindow).

frameDefaultStyle :: Style

The default frame style for a normal top-level Frame.

frameCenter :: Frame a -> IO ()

Center the frame on the screen.

frameCenterHorizontal :: Frame a -> IO ()

Center the frame horizontally on the screen.

frameCenterVertical :: Frame a -> IO ()

Center the frame vertically on the screen.

Window

windowGetRootParent :: Window a -> IO (Window ())

The ultimate root parent of the widget.

windowGetFrameParent :: Window a -> IO (Window ())

The parent frame or dialog of a widget.

windowGetMousePosition :: Window a -> IO Point

Retrieve the current mouse position relative to the window position.

windowGetScreenPosition :: Window a -> IO Point

Get the window position relative to the origin of the display.

windowChildren :: Window a -> IO [Window ()]

Get the children of a window

Dialog

dialogDefaultStyle :: Style

The default frame style for a normal Dialog.

Status bar

statusBarCreateFields :: Frame a -> [Int] -> IO (StatusBar ())