Graphics.UI.WXCore.WxcClassInfo

Description

Haskell class info definitions for the wxWidgets C library (wxc.dll).

Do not edit this file manually! This file was automatically generated by wxDirect.

And contains 406 class info definitions.

Class Info

data ClassType a

The type of a class.

classInfo :: ClassType a -> ClassInfo ()

Return the ClassInfo belonging to a class type. (Do not delete this object, it is statically allocated)

instanceOf :: WxObject b -> ClassType a -> Bool

Test if an object is of a certain kind. (Returns also True when the object is null.)

instanceOfName :: WxObject a -> String -> Bool

Test if an object is of a certain kind, based on a full wxWidgets class name. (Use with care).

Safe casts

safeCast :: WxObject b -> ClassType (WxObject a) -> Maybe (WxObject a)

A safe object cast. Returns Nothing if the object is of the wrong type. Note that a null object can always be cast.

ifInstanceOf :: WxObject a -> ClassType (WxObject b) -> (WxObject b -> c) -> c -> c

Perform an action when the object has the right kind. Perform the default action if the kind is not correct. Note that a null object has always the right kind.

whenInstanceOf :: WxObject a -> ClassType (WxObject b) -> (WxObject b -> IO ()) -> IO ()

Perform an action when the object has the right kind. Note that a null object has always the right kind.

whenValidInstanceOf :: WxObject a -> ClassType (WxObject b) -> (WxObject b -> IO ()) -> IO ()

Perform an action when the object has the right type and is not null.

Class Types

Down casts