Graphics.UI.WXCore.WxcDefs

Description

Haskell constant definitions for the wxWidgets C library (wxc.dll).

This file was originally generated automatically by wxDirect; it is now manually maintained.

Types

type BitFlag = Int

A flag can be combined with other flags to a bit mask.

Constants

wxBORDER_DEFAULT :: Int

This is different from wxBORDER_NONE, as by default the controls do have border

wxBORDER_MASK :: Int

A mask to extract border style from the combination of flags

wxBORDER_NONE :: Int

Displays no border, overriding the default border style for the window. wxNO_BORDER is the old name for this style.

wxBORDER_RAISED :: Int

Displays a raised border. wxRAISED_BORDER is the old name for this style.

wxBORDER_SIMPLE :: Int

Displays a thin border around the window. wxSIMPLE_BORDER is the old name for this style.

wxBORDER_STATIC :: Int

Displays a border suitable for a static control. wxSTATIC_BORDER is the old name for this style. Windows only.

wxBORDER_SUNKEN :: Int

Displays a sunken border. wxSUNKEN_BORDER is the old name for this style.

wxBORDER_THEME :: Int

Displays a themed border where possible. Currently this has an effect on Windows XP and above only. For more information on themed borders, please see Themed borders on Windows http://docs.wxwidgets.org/2.8/wx_wxmswport.html#wxmswthemedborders.

wxDOUBLE_BORDER :: Int

Deprecated: Use wxBORDER_THEME

Deprecated

wxRAISED_BORDER :: Int

Deprecated: Use wxBORDER_RAISED

Deprecated

wxSTATIC_BORDER :: Int

Deprecated: Use wxBORDER_STATIC

Deprecated

wxSUNKEN_BORDER :: Int

Deprecated: Use wxBORDER_SUNKEN

Deprecated